refactor(#742): relocate built-in static-method dispatch to call-builtin-static.ts
Wave B chunk 2, step 2. Move compileBuiltinStaticCall (Math / BigInt / Number /
Array / String / Object namespace statics) out of calls.ts into a new sibling
module src/codegen/expressions/call-builtin-static.ts. Nine calls.ts internals
are exported to it; calls.ts imports compileBuiltinStaticCall back (lazy cycle,
same pattern as call-identifier.ts / calls-closures.ts).

calls.ts: 17,462 -> 14,484 LOC. compileCallExpression: ~11,388 -> ~8,410 LOC.
New call-builtin-static.ts: 3,064 LOC.

Pure relocation — emitted Wasm byte-identical (prove-emit-identity IDENTICAL,
39/39 gc/standalone/wasi). tsc 0 errors. Smoke test tests/issue-742.test.ts
extended with Math/Number/Array/String/Object static-call cases. The
loc-budget-allow key in the issue frontmatter sanctions the new file the
change-scoped loc-budget gate flags (net-zero relocation); coercion-sites and
oracle-ratchet pass with no allowance.

✓
